+++ threedi_modelchecker/config.py
@@ -61,6 +61,7 @@ from .checks.other import (
     DefinedAreaCheck,
     FeatureClosedCrossSectionCheck,
     ImperviousNodeInflowAreaCheck,
+    InflowNoFeaturesCheck,
     LinestringLocationCheck,
     NodeSurfaceConnectionsCheck,
     OpenChannelsWithNestedNewton,
@@ -2070,7 +2071,33 @@ CHECKS += [
         (models.ImperviousSurfaceMap.connection_node_id, models.ConnectionNode),
     )
 ]
+CHECKS += [
+    QueryCheck(
+        error_code=616,
+        level=CheckLevel.WARNING,
+        column=table.id,
+        filters=~CONDITIONS[condition].exists(),
+        invalid=Query(table),
+        message=f"No inflow will be generated for this feature, because v2_global_settings.use_0d_inflow is not set to use {table.__tablename__}.",
+    )
+    for table, condition in (
+        (models.ImperviousSurface, "0d_imp"),
+        (models.Surface, "0d_surf"),
+    )
+]
 
+CHECKS += [
+    InflowNoFeaturesCheck(
+        error_code=617,
+        level=CheckLevel.WARNING,
+        surface_table=table,
+        filters=CONDITIONS[condition].exists(),
+    )
+    for table, condition in (
+        (models.ImperviousSurface, "0d_imp"),
+        (models.Surface, "0d_surf"),
+    )
+]
 
 CHECKS += [
     RangeCheck(