+++ CHANGES.rst
@@ -5,7 +5,8 @@ Changelog of threedi-modelchecker
 2.6.2 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Add warning check (0616) for surfaces for which no inflow is generated because of the surface table not being referred to in global settings.
+- Add warning check (0617) to warn if the surface table referred to in global settings is empty and no inflow will be generated for it.
 
 
 2.6.1 (2024-02-20)